vidstige/jadb#26 separate tests for unit and integration
