+++ pom.xml
@@ -27,6 +27,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <maven.compiler.target>1.7</maven.compiler.target>
         <maven.compiler.source>1.7</maven.compiler.source>
     </properties>
@@ -82,18 +83,44 @@
                 </dependencies>
             </plugin>
             <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.5</version>
+                <version>2.19.1</version>
                 <configuration>
                     <includes>
                         <include>**/*.java</include>
                     </includes>
                     <excludes>
+                        <!--Integration tests -->
+                        <exclude>se.vidstige.jadb.test.integration.*</exclude>
+                        <!--Mocks and data-->
                         <exclude>**/data/*</exclude>
                         <exclude>**/fakes/*</exclude>
                     </excludes>
                 </configuration>
+            </plugin>
+            <plugin>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <version>2.19.1</version>
+                <configuration>
+                    <includes>
+                        <include>**/*.java</include>
+                    </includes>
+                    <excludes>
+                        <!--Integration tests -->
+                        <exclude>se.vidstige.jadb.test.unit.*</exclude>
+                        <!--Mocks and data-->
+                        <exclude>**/data/*</exclude>
+                        <exclude>**/fakes/*</exclude>
+                    </excludes>
+                </configuration>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>integration-test</goal>
+                            <goal>verify</goal>
+                        </goals>
+                    </execution>
+                </executions>
             </plugin>
         </plugins>
     </build>
